Egaz support added

+++ controllers/corpus.js
@@ -14,76 +14,115 @@ const translate = new Translate({
 });
 const target = 'eng';
 
-exports.processImage = function (req, res) {
-    let imagePath = req.imagePath
-    let file_base_name = imagePath.replace('.png', '')
-    exec('tesseract ' + imagePath + ' ' + file_base_name + ' -l hin+eng', (err, stdout, stderr) => {
+exports.processImage = function (req, res, dontSendRes) {
+    let imagePaths = req.image_paths
+    let file_base_name = imagePaths[0].replace('.png', '').split('-')[0]
+    callTesseractForMultipleLanguage(imagePaths, 0, res, file_base_name, dontSendRes)
+}
+
+
+function callTesseractForMultipleLanguage(imagePaths, index, res, file_base_name, dontSendRes) {
+    exec('tesseract ' + imagePaths[index] + ' - >> ' + file_base_name + '.txt -l hin+eng', (err, stdout, stderr) => {
+        index++;
         if (err) {
-            let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
-            return res.status(apistatus.http.status).json(apistatus);
-        }
-        var exec_cmd = python_version + ' separate.py ' + file_base_name + '.txt ' + file_base_name
-        exec(exec_cmd, (err, stdout, stderr) => {
-            if (err) {
+            console.log(err)
+            if (!dontSendRes) {
                 let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
                 return res.status(apistatus.http.status).json(apistatus);
             }
-            var exec_cmd = python_version + ' ' + 'process_paragraph.py' + ' ' + file_base_name + '_hin.txt ' + file_base_name
+        }
+
+        if (index == imagePaths.length) {
+            var exec_cmd = python_version + ' separate.py ' + file_base_name + '.txt ' + file_base_name
             exec(exec_cmd, (err, stdout, stderr) => {
                 if (err) {
-                    let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
-                    return res.status(apistatus.http.status).json(apistatus);
-                }
-                var exec_cmd = python_version + ' ' + 'process_paragraph_eng.py' + ' ' + file_base_name + '_eng.txt ' + file_base_name
-                exec(exec_cmd, (err, stdout, stderr) => {
-                    if (err) {
+                    if (!dontSendRes) {
                         let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
                         return res.status(apistatus.http.status).json(apistatus);
                     }
-
-                    fs.readFile(file_base_name + '_hin' + '.txt', 'utf8', function (err, data) {
-                        if (err) {
+                }
+                var exec_cmd = python_version + ' ' + 'process_paragraph.py' + ' ' + file_base_name + '_hin.txt ' + file_base_name
+                exec(exec_cmd, (err, stdout, stderr) => {
+                    if (err) {
+                        if (!dontSendRes) {
                             let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
                             return res.status(apistatus.http.status).json(apistatus);
                         }
-                        translate
-                            .translate(data, target)
-                            .then(results => {
-                                const translation = results[0];
-                                fs.writeFile(file_base_name + '_eng_tran' + '.txt', translation, function (err) {
-                                    if (err) {
-                                        let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
-                                        return res.status(apistatus.http.status).json(apistatus);
-                                    }
-                                    let corpus_cmd = './helpers/bleualign.py -s ' + __dirname + '/../' + file_base_name + '_hin' + '.txt' + ' -t ' + __dirname + '/../' + file_base_name + '_eng' + '.txt' + ' --srctotarget ' + __dirname + '/../' + file_base_name + '_eng_tran' + '.txt' + ' -o ' + __dirname + '/../' + file_base_name + '_output'
-                                    exec(corpus_cmd, (err, stdout, stderr) => {
-                                        if (err) {
-                                            let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
-                                            return res.status(apistatus.http.status).json(apistatus);
-                                        }
-                                        let output_data = {}
-                                        fs.readFile(file_base_name + '_output-s', 'utf8', function (err, data) {
-                                            output_data.hindi = data.split('\n')
-                                            fs.readFile(file_base_name + '_output-t', 'utf8', function (err, data) {
-                                                output_data.english = data.split('\n')
-                                                glob(file_base_name + "*", function (er, files) {
-                                                    if (files && files.length > 0) {
-                                                        files.map((fileName) => {
-                                                            fs.unlink(fileName, function () { })
-                                                        })
+                    }
+                    var exec_cmd = python_version + ' ' + 'process_paragraph_eng.py' + ' ' + file_base_name + '_eng.txt ' + file_base_name
+                    exec(exec_cmd, (err, stdout, stderr) => {
+                        if (err) {
+                            if (!dontSendRes) {
+                                let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
+                                return res.status(apistatus.http.status).json(apistatus);
+                            }
+                        }
+
+                        fs.readFile(file_base_name + '_hin' + '.txt', 'utf8', function (err, data) {
+                            if (err) {
+                                if (!dontSendRes) {
+                                    let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
+                                    return res.status(apistatus.http.status).json(apistatus);
+                                }
+                            }
+                            let data_arr = data.split('\n')
+                            if (data_arr.length > 100) {
+                                let loops = Math.ceil(data_arr.length / 100)
+                                let translated_text = ''
+                                transalteBigText(0, loops, data_arr, res, translated_text, file_base_name, dontSendRes)
+                            }
+                            else {
+                                translate
+                                    .translate(data, target)
+                                    .then(results => {
+                                        const translation = results[0];
+                                        fs.writeFile(file_base_name + '_eng_tran' + '.txt', translation, function (err) {
+                                            if (err) {
+                                                if (!dontSendRes) {
+                                                    let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
+                                                    return res.status(apistatus.http.status).json(apistatus);
+                                                }
+                                            }
+                                            let output_file_base = new Date().getTime()
+                                            let corpus_cmd = './helpers/bleualign.py -s ' + __dirname + '/../' + file_base_name + '_hin' + '.txt' + ' -t ' + __dirname + '/../' + file_base_name + '_eng' + '.txt' + ' --srctotarget ' + __dirname + '/../' + file_base_name + '_eng_tran' + '.txt' + ' -o ' + __dirname + '/../upload/' + output_file_base + '_output'
+                                            exec(corpus_cmd, (err, stdout, stderr) => {
+                                                if (err) {
+                                                    if (!dontSendRes) {
+                                                        let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
+                                                        return res.status(apistatus.http.status).json(apistatus);
                                                     }
-                                                })
-                                                let apistatus = new Response(StatusCode.SUCCESS, output_data).getRsp()
-                                                return res.status(apistatus.http.status).json(apistatus);
+                                                }
+                                                let output_data = {}
+                                                fs.readFile('upload/' + output_file_base + '_output' + '-s', 'utf8', function (err, data) {
+                                                    output_data.hindi = data.split('\n')
+                                                    fs.readFile('upload/' + output_file_base + '_output' + '-t', 'utf8', function (err, data) {
+                                                        output_data.english = data.split('\n')
+                                                        glob(file_base_name + "*", function (er, files) {
+                                                            if (files && files.length > 0) {
+                                                                files.map((fileName) => {
+                                                                    if (fileName.indexOf('pdf') == -1)
+                                                                        fs.unlink(fileName, function () { })
+                                                                })
+                                                            }
+                                                        })
+                                                        if (!dontSendRes) {
+                                                            let apistatus = new Response(StatusCode.SUCCESS, output_data).getRsp()
+                                                            return res.status(apistatus.http.status).json(apistatus);
+                                                        }
+                                                    })
+                                                });
                                             })
                                         });
                                     })
-                                });
-                            })
+                            }
+                        })
                     })
-                })
+                });
             });
-        });
+        }
+        else {
+            callTesseractForMultipleLanguage(imagePaths, index, res, file_base_name, dontSendRes)
+        }
     });
 
 }
@@ -155,7 +194,7 @@ exports.convertAndCreateCorpus = function (req, res) {
 }
 
 
-function transalteBigText(i, loops, data_arr, res, translated_text, file_base_name) {
+function transalteBigText(i, loops, data_arr, res, translated_text, file_base_name, dontSendRes) {
     let endCount = 100 > data_arr.length ? data_arr.length % 100 : 100
     translate
         .translate(data_arr.splice(0, endCount), target)
@@ -167,8 +206,10 @@ function transalteBigText(i, loops, data_arr, res, translated_text, file_base_na
             if (i + 1 == loops) {
                 fs.writeFile(file_base_name + '_eng_tran' + '.txt', translated_text, function (err) {
                     if (err) {
-                        let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
-                        return res.status(apistatus.http.status).json(apistatus);
+                        if (!dontSendRes) {
+                            let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
+                            return res.status(apistatus.http.status).json(apistatus);
+                        }
                     }
                     let output_file_base = new Date().getTime()
                     let corpus_cmd = './helpers/bleualign.py -s ' + __dirname + '/../' + file_base_name + '_hin' + '.txt' + ' -t ' + __dirname + '/../' + file_base_name + '_eng' + '.txt' + ' --srctotarget ' + __dirname + '/../' + file_base_name + '_eng_tran' + '.txt' + ' -o ' + __dirname + '/../upload/' + output_file_base + '_output'
@@ -176,8 +217,10 @@ function transalteBigText(i, loops, data_arr, res, translated_text, file_base_na
                     exec(corpus_cmd, (err, stdout, stderr) => {
                         if (err) {
                             console.log(err)
-                            let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
-                            return res.status(apistatus.http.status).json(apistatus);
+                            if (!dontSendRes) {
+                                let apistatus = new APIStatus(StatusCode.ERR_GLOBAL_SYSTEM, 'app').getRspStatus()
+                                return res.status(apistatus.http.status).json(apistatus);
+                            }
                         }
                         let output_data = {}
                         fs.readFile('upload/' + output_file_base + '_output' + '-s', 'utf8', function (err, data) {
@@ -187,12 +230,15 @@ function transalteBigText(i, loops, data_arr, res, translated_text, file_base_na
                                 glob(file_base_name + "*", function (er, files) {
                                     if (files && files.length > 0) {
                                         files.map((fileName) => {
-                                            fs.unlink(fileName, function () { })
+                                            if (fileName.indexOf('pdf') == -1)
+                                                fs.unlink(fileName, function () { })
                                         })
                                     }
                                 })
-                                let apistatus = new Response(StatusCode.SUCCESS, output_data).getRsp()
-                                return res.status(apistatus.http.status).json(apistatus);
+                                if (!dontSendRes) {
+                                    let apistatus = new Response(StatusCode.SUCCESS, output_data).getRsp()
+                                    return res.status(apistatus.http.status).json(apistatus);
+                                }
                             })
                         });
                     })
@@ -200,7 +246,7 @@ function transalteBigText(i, loops, data_arr, res, translated_text, file_base_na
             }
             else {
                 i = i + 1;
-                transalteBigText(i, loops, data_arr, res, translated_text, file_base_name)
+                transalteBigText(i, loops, data_arr, res, translated_text, file_base_name, dontSendRes)
             }
         }).catch((e) => {
             console.log(e)
